requirements.txt
-i https://pypi.org/simple/
click==7.1.2
flask==1.0.4
itsdangerous==1.1.0
jinja2==2.11.3
markupsafe==1.1.1
pillow==6.2.0
pytesseract==0.2.6
werkzeug==1.0.1
